feat: SpendEntity 생성 및 Query DSL 연동, SpendSpecification page,size type 변경

+++ budget-domain/src/main/java/com/hg/budget/domain/spend/SpendService.java
@@ -54,8 +54,8 @@ public class SpendService {
     }
 
     public Page<Spend> findSpendList(
-        long page,
-        long size,
+        int page,
+        int size,
         LocalDateTime startSpentDateTime,
         LocalDateTime endSpentDateTime,
         Account account,
